Move HydrogenResponse to foundation
