+++ packages/hydrogen/src/utilities/log/log-cache-header.ts
@@ -1,5 +1,5 @@
 import {HydrogenRequest} from '../../foundation/HydrogenRequest/HydrogenRequest.server';
-import {HydrogenResponse} from '../../framework/HydrogenResponse.server';
+import {HydrogenResponse} from '../../foundation/HydrogenResponse/HydrogenResponse.server';
 import {QueryKey} from '../../types';
 import {hashKey} from '../hash';
 import {findQueryName, parseUrl} from './utils';
